requirements.txt
grpcio
grpcio-tools
pika
loguru